Added obj fun history and details to Solution schema

+++ flowing_basin/solvers/lp.py
@@ -24,7 +24,7 @@ class LPConfiguration:
     MIPGap: float
 
     # Solver timeout
-    time_limit_seconds: int
+    time_limit_seconds: float
 
 
 class LPModel(Experiment):
